Take size_t and cast inside, since most things are going to pass size_t so might as well jsut cast in one place.

+++ Pulp/3DEngine/Drawing/VariableStateManager.h
@@ -40,7 +40,7 @@ public:
 	~VariableStateManager();
 
 
-	render::Commands::ResourceStateBase* createVariableState(int8_t numTexStates, int8_t numCBs);
+	render::Commands::ResourceStateBase* createVariableState(size_t numTexStates, size_t numCBs);
 
 private:
 	render::Commands::ResourceStateBase* createVariableState_Interal(int8_t numTexStates, int8_t numCBs);
